updacte readme for agent refactor + static manipulator abstract away
