+++ habitat-lab/habitat/articulated_agents/robots/franka_robot.py
@@ -5,7 +5,7 @@
 import magnum as mn
 import numpy as np
 
-from habitat.articulated_agents.robots.static_manipulator import (
+from habitat.articulated_agents.static_manipulator import (
     StaticManipulator,
     StaticManipulatorParams,
 )
