+++ habitat-lab/habitat/articulated_agents/static_manipulator.py
